app/requirements.txt
python-magic
pickle5
bs4
Pillow
SpeechRecognition
scikit-learn
python-docx